refactor: extract shared FilterComponents from example pages
Move renderValueInput, ConditionRow, and GroupedFilters into a shared
FilterComponents module typed generically over TSchema. Both pages now
import from the same component set instead of duplicating ~200 lines
of identical logic.
